+++ packages/api/src/main.rs
@@ -13,7 +13,7 @@ use models::{
     response::SurveyResponse,
     v2::{Institution, PublicOpinionProject},
 };
-use models::{v2::Review, *};
+use models::{step::Step, v2::Review, *};
 use sqlx::postgres::PgPoolOptions;
 // use by_types::DatabaseConfig;
 // use sqlx::postgres::PgPoolOptions;
